pyproject.toml
[tool.poetry]
name = "smack"
version = "0.1.0"
description = ""
authors = ["Max Wolschlager <rxtx@ungut.at>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
rich = "^13.9.3"
pyfiglet = "^1.0.2"
typer = "^0.12.5"
pyyaml = "^6.0.2"
pynput = "^1.7.7"
markdown-it-py = {extras = ["plugins"], version = "^3.0.0"}
mdformat = "^0.7.18"
polars = "^1.12.0"
termplotlib = "^0.3.9"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
ruff = "^0.7.1"
types-pyyaml = "^6.0.12.20240917"

[tool.poetry.scripts]
smack = "smack.__main__:app"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
exclude = [
]
disable_error_code = ['override']
implicit_reexport = true

[[tool.mypy.overrides]]
module = ["marko.*", "pyfiglet", "termplotlib"]
ignore_missing_imports = true
